Struct aws_sdk_neptune::client::fluent_builders::CreateDBInstance

pub struct CreateDBInstance<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateDBInstance<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateDbInstanceOutput, SdkError<CreateDBInstanceError>> where
R::Policy: SmithyRetryPolicy<CreateDbInstanceInputOperationOutputAlias, CreateDbInstanceOutput, CreateDBInstanceError, CreateDbInstanceInputOperationRetryAlias>,

pub fn db_name(self, inp: impl Into<String>) -> Self

Not supported.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn db_instance_identifier(self, inp: impl Into<String>) -> Self

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn allocated_storage(self, inp: i32) -> Self

Not supported by Neptune.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn db_instance_class(self, inp: impl Into<String>) -> Self

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Regions.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn engine(self, inp: impl Into<String>) -> Self

The name of the database engine to be used for this instance.

Valid Values: neptune

pub fn set_engine(self, input: Option<String>) -> Self

pub fn master_username(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn master_user_password(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn db_security_groups(self, inp: impl Into<String>) -> Self

Appends an item to DBSecurityGroups.

To override the contents of this collection use set_db_security_groups.

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

A list of EC2 VPC security groups to associate with this DB instance.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn availability_zone(self, inp: impl Into<String>) -> Self

The EC2 Availability Zone that the DB instance is created in

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same Amazon Region as the current endpoint.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, inp: impl Into<String>) -> Self

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, inp: i32) -> Self

The number of days for which automated backups are retained.

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn preferred_backup_window(self, inp: impl Into<String>) -> Self

The daily time range during which automated backups are created.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which the database accepts connections.

Not applicable. The port is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 8182

Type: Integer

pub fn set_port(self, input: Option<i32>) -> Self

pub fn multi_az(self, inp: bool) -> Self

Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version number of the database engine to use. Currently, setting this parameter has no effect.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, inp: impl Into<String>) -> Self

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, inp: i32) -> Self

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, inp: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn character_set_name(self, inp: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_character_set_name(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, inp: bool) -> Self

This flag should no longer be used.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

The tags to assign to the new instance.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn db_cluster_identifier(self, inp: impl Into<String>) -> Self

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn storage_type(self, inp: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Not applicable. Storage is managed by the DB Cluster.

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, inp: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn tde_credential_password(self, inp: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, inp: bool) -> Self

Specifies whether the DB instance is encrypted.

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

The Amazon KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same Amazon account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon Neptune will use your default encryption key. Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn domain(self, inp: impl Into<String>) -> Self

Specify the Active Directory Domain to create the instance in.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, inp: bool) -> Self

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, inp: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn monitoring_role_arn(self, inp: impl Into<String>) -> Self

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, inp: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn promotion_tier(self, inp: i32) -> Self

A value that specifies the order in which an Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

pub fn timezone(self, inp: impl Into<String>) -> Self

The time zone of the DB instance.

pub fn set_timezone(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, inp: bool) -> Self

Not supported by Neptune (ignored).

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_performance_insights(self, inp: bool) -> Self

(Not supported by Neptune)

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, inp: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

Appends an item to EnableCloudwatchLogsExports.

To override the contents of this collection use set_enable_cloudwatch_logs_exports.

The list of log types that need to be enabled for exporting to CloudWatch Logs.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn deletion_protection(self, inp: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

DB instances in a DB cluster can be deleted even when deletion protection is enabled in their parent DB cluster.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self