Struct aws_sdk_neptune::client::fluent_builders::ModifyDBCluster

pub struct ModifyDBCluster<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> ModifyDBCluster<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<ModifyDbClusterOutput, SdkError<ModifyDBClusterError>> where
R::Policy: SmithyRetryPolicy<ModifyDbClusterInputOperationOutputAlias, ModifyDbClusterOutput, ModifyDBClusterError, ModifyDbClusterInputOperationRetryAlias>,

pub fn db_cluster_identifier(self, inp: impl Into<String>) -> Self

The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn new_db_cluster_identifier(self, inp: impl Into<String>) -> Self

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

pub fn set_new_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn apply_immediately(self, inp: bool) -> Self

A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects NewDBClusterIdentifier values. If you set the ApplyImmediately parameter value to false, then changes to NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn backup_retention_period(self, inp: i32) -> Self

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn db_cluster_parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the DB cluster parameter group to use for the DB cluster.

pub fn set_db_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

A list of VPC security groups that the DB cluster will belong to.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_user_password(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn option_group_name(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, inp: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, inp: bool) -> Self

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn cloudwatch_logs_export_configuration(
self,
inp: CloudwatchLogsExportConfiguration
) -> Self

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

pub fn set_cloudwatch_logs_export_configuration(
self,
input: Option<CloudwatchLogsExportConfiguration>
) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true.

For a list of valid engine versions, see Engine Releases for Amazon Neptune, or call DescribeDBEngineVersions.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn deletion_protection(self, inp: bool) -> Self

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn copy_tags_to_snapshot(self, inp: bool) -> Self

If set to true, tags are copied to any snapshot of the DB cluster that is created.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self