Struct aws_sdk_neptune::client::fluent_builders::ModifyDBInstance

pub struct ModifyDBInstance<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> ModifyDBInstance<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<ModifyDbInstanceOutput, SdkError<ModifyDBInstanceError>> where
R::Policy: SmithyRetryPolicy<ModifyDbInstanceInputOperationOutputAlias, ModifyDbInstanceOutput, ModifyDBInstanceError, ModifyDbInstanceInputOperationRetryAlias>,

pub fn db_instance_identifier(self, inp: impl Into<String>) -> Self

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn allocated_storage(self, inp: i32) -> Self

Not supported by Neptune.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn db_instance_class(self, inp: impl Into<String>) -> Self

The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, inp: impl Into<String>) -> Self

The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you specify true for the ApplyImmediately parameter.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn db_security_groups(self, inp: impl Into<String>) -> Self

Appends an item to DBSecurityGroups.

To override the contents of this collection use set_db_security_groups.

A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn apply_immediately(self, inp: bool) -> Self

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance.

If this parameter is set to false, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot.

Default: false

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn master_user_password(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. The db instance will NOT be rebooted automatically and the parameter changes will NOT be applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, inp: i32) -> Self

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn preferred_backup_window(self, inp: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn multi_az(self, inp: bool) -> Self

Specifies if the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version number of the database engine to upgrade to. Currently, setting this parameter has no effect. To upgrade your database engine to the most recent release, use the ApplyPendingMaintenanceAction API.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn allow_major_version_upgrade(self, inp: bool) -> Self

Indicates that major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

pub fn set_allow_major_version_upgrade(self, input: Option<bool>) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

Indicates that minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and Neptune has enabled auto patching for that engine version.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, inp: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, inp: i32) -> Self

The new Provisioned IOPS (I/O operations per second) value for the instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

Default: Uses existing setting

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, inp: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn new_db_instance_identifier(self, inp: impl Into<String>) -> Self

The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot will occur immediately if you set Apply Immediately to true, or will occur during the next maintenance window if Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

pub fn set_new_db_instance_identifier(self, input: Option<String>) -> Self

pub fn storage_type(self, inp: impl Into<String>) -> Self

Not supported.

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, inp: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn tde_credential_password(self, inp: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

pub fn ca_certificate_identifier(self, inp: impl Into<String>) -> Self

Indicates the certificate that needs to be associated with the instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

pub fn domain(self, inp: impl Into<String>) -> Self

Not supported.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, inp: bool) -> Self

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, inp: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn db_port_number(self, inp: i32) -> Self

The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

Default: 8182

pub fn set_db_port_number(self, input: Option<i32>) -> Self

pub fn publicly_accessible(self, inp: bool) -> Self

This flag should no longer be used.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn monitoring_role_arn(self, inp: impl Into<String>) -> Self

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, inp: impl Into<String>) -> Self

Not supported

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn promotion_tier(self, inp: i32) -> Self

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

pub fn enable_iam_database_authentication(self, inp: bool) -> Self

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines

Not applicable. Mapping Amazon IAM accounts to database accounts is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: false

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_performance_insights(self, inp: bool) -> Self

(Not supported by Neptune)

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, inp: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn cloudwatch_logs_export_configuration(
self,
inp: CloudwatchLogsExportConfiguration
) -> Self

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance or DB cluster.

pub fn set_cloudwatch_logs_export_configuration(
self,
input: Option<CloudwatchLogsExportConfiguration>
) -> Self

pub fn deletion_protection(self, inp: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self