Module aws_sdk_neptune::error

Modules

authorization_not_found_fault

See AuthorizationNotFoundFault

certificate_not_found_fault

See CertificateNotFoundFault

db_cluster_already_exists_fault

See DbClusterAlreadyExistsFault

db_cluster_endpoint_already_exists_fault

See DbClusterEndpointAlreadyExistsFault

db_cluster_endpoint_not_found_fault

See DbClusterEndpointNotFoundFault

db_cluster_endpoint_quota_exceeded_fault

See DbClusterEndpointQuotaExceededFault

db_cluster_not_found_fault

See DbClusterNotFoundFault

db_cluster_parameter_group_not_found_fault

See DbClusterParameterGroupNotFoundFault

db_cluster_quota_exceeded_fault

See DbClusterQuotaExceededFault

db_cluster_role_already_exists_fault

See DbClusterRoleAlreadyExistsFault

db_cluster_role_not_found_fault

See DbClusterRoleNotFoundFault

db_cluster_role_quota_exceeded_fault

See DbClusterRoleQuotaExceededFault

db_cluster_snapshot_already_exists_fault

See DbClusterSnapshotAlreadyExistsFault

db_cluster_snapshot_not_found_fault

See DbClusterSnapshotNotFoundFault

db_instance_already_exists_fault

See DbInstanceAlreadyExistsFault

db_instance_not_found_fault

See DbInstanceNotFoundFault

db_parameter_group_already_exists_fault

See DbParameterGroupAlreadyExistsFault

db_parameter_group_not_found_fault

See DbParameterGroupNotFoundFault

db_parameter_group_quota_exceeded_fault

See DbParameterGroupQuotaExceededFault

db_security_group_not_found_fault

See DbSecurityGroupNotFoundFault

db_snapshot_already_exists_fault

See DbSnapshotAlreadyExistsFault

db_snapshot_not_found_fault

See DbSnapshotNotFoundFault

db_subnet_group_already_exists_fault

See DbSubnetGroupAlreadyExistsFault

db_subnet_group_does_not_cover_enough_a_zs

See DbSubnetGroupDoesNotCoverEnoughAZs

db_subnet_group_not_found_fault

See DbSubnetGroupNotFoundFault

db_subnet_group_quota_exceeded_fault

See DbSubnetGroupQuotaExceededFault

db_subnet_quota_exceeded_fault

See DbSubnetQuotaExceededFault

db_upgrade_dependency_failure_fault

See DbUpgradeDependencyFailureFault

domain_not_found_fault

See DomainNotFoundFault

event_subscription_quota_exceeded_fault

See EventSubscriptionQuotaExceededFault

instance_quota_exceeded_fault

See InstanceQuotaExceededFault

insufficient_db_cluster_capacity_fault

See InsufficientDbClusterCapacityFault

insufficient_db_instance_capacity_fault

See InsufficientDbInstanceCapacityFault

insufficient_storage_cluster_capacity_fault

See InsufficientStorageClusterCapacityFault

invalid_db_cluster_endpoint_state_fault

See InvalidDbClusterEndpointStateFault

invalid_db_cluster_snapshot_state_fault

See InvalidDbClusterSnapshotStateFault

invalid_db_cluster_state_fault

See InvalidDbClusterStateFault

invalid_db_instance_state_fault

See InvalidDbInstanceStateFault

invalid_db_parameter_group_state_fault

See InvalidDbParameterGroupStateFault

invalid_db_security_group_state_fault

See InvalidDbSecurityGroupStateFault

invalid_db_snapshot_state_fault

See InvalidDbSnapshotStateFault

invalid_db_subnet_group_state_fault

See InvalidDbSubnetGroupStateFault

invalid_db_subnet_state_fault

See InvalidDbSubnetStateFault

invalid_event_subscription_state_fault

See InvalidEventSubscriptionStateFault

invalid_restore_fault

See InvalidRestoreFault

invalid_subnet

See InvalidSubnet

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

kms_key_not_accessible_fault

See KmsKeyNotAccessibleFault

option_group_not_found_fault

See OptionGroupNotFoundFault

provisioned_iops_not_available_in_az_fault

See ProvisionedIopsNotAvailableInAzFault

resource_not_found_fault

See ResourceNotFoundFault

shared_snapshot_quota_exceeded_fault

See SharedSnapshotQuotaExceededFault

snapshot_quota_exceeded_fault

See SnapshotQuotaExceededFault

sns_invalid_topic_fault

See SnsInvalidTopicFault

sns_no_authorization_fault

See SnsNoAuthorizationFault

sns_topic_arn_not_found_fault

See SnsTopicArnNotFoundFault

source_not_found_fault

See SourceNotFoundFault

storage_quota_exceeded_fault

See StorageQuotaExceededFault

storage_type_not_supported_fault

See StorageTypeNotSupportedFault

subnet_already_in_use

See SubnetAlreadyInUse

subscription_already_exist_fault

See SubscriptionAlreadyExistFault

subscription_category_not_found_fault

See SubscriptionCategoryNotFoundFault

subscription_not_found_fault

See SubscriptionNotFoundFault

Structs

AddRoleToDBClusterError

AddSourceIdentifierToSubscriptionError

AddTagsToResourceError

ApplyPendingMaintenanceActionError

AuthorizationNotFoundFault

Specified CIDRIP or EC2 security group is not authorized for the specified DB security group.

CertificateNotFoundFault

CopyDBClusterParameterGroupError

CopyDBClusterSnapshotError

CopyDBParameterGroupError

CreateDBClusterEndpointError

CreateDBClusterError

CreateDBClusterParameterGroupError

CreateDBClusterSnapshotError

CreateDBInstanceError

CreateDBParameterGroupError

CreateDBSubnetGroupError

CreateEventSubscriptionError

DbClusterAlreadyExistsFault

User already has a DB cluster with the given identifier.

DbClusterEndpointAlreadyExistsFault

The specified custom endpoint cannot be created because it already exists.

DbClusterEndpointNotFoundFault

The specified custom endpoint doesn't exist.

DbClusterEndpointQuotaExceededFault

The cluster already has the maximum number of custom endpoints.

DbClusterNotFoundFault

DbClusterParameterGroupNotFoundFault

DbClusterQuotaExceededFault

User attempted to create a new DB cluster and the user has already reached the maximum allowed DB cluster quota.

DbClusterRoleAlreadyExistsFault

The specified IAM role Amazon Resource Name (ARN) is already associated with the specified DB cluster.

DbClusterRoleNotFoundFault

The specified IAM role Amazon Resource Name (ARN) is not associated with the specified DB cluster.

DbClusterRoleQuotaExceededFault

You have exceeded the maximum number of IAM roles that can be associated with the specified DB cluster.

DbClusterSnapshotAlreadyExistsFault

User already has a DB cluster snapshot with the given identifier.

DbClusterSnapshotNotFoundFault

DbInstanceAlreadyExistsFault

User already has a DB instance with the given identifier.

DbInstanceNotFoundFault

DbParameterGroupAlreadyExistsFault

A DB parameter group with the same name exists.

DbParameterGroupNotFoundFault

DbParameterGroupQuotaExceededFault

Request would result in user exceeding the allowed number of DB parameter groups.

DbSecurityGroupNotFoundFault

DbSnapshotAlreadyExistsFault

DbSnapshotNotFoundFault

DbSubnetGroupAlreadyExistsFault

DbSubnetGroupDoesNotCoverEnoughAZs

Subnets in the DB subnet group should cover at least two Availability

DbSubnetGroupNotFoundFault

DbSubnetGroupQuotaExceededFault

Request would result in user exceeding the allowed number of DB subnet groups.

DbSubnetQuotaExceededFault

Request would result in user exceeding the allowed number of subnets in a DB subnet groups.

DbUpgradeDependencyFailureFault

The DB upgrade failed because a resource the DB depends on could not be modified.

DeleteDBClusterEndpointError

DeleteDBClusterError

DeleteDBClusterParameterGroupError

DeleteDBClusterSnapshotError

DeleteDBInstanceError

DeleteDBParameterGroupError

DeleteDBSubnetGroupError

DeleteEventSubscriptionError

DescribeDBClusterEndpointsError

DescribeDBClusterParameterGroupsError

DescribeDBClusterParametersError

DescribeDBClusterSnapshotAttributesError

DescribeDBClusterSnapshotsError

DescribeDBClustersError

DescribeDBEngineVersionsError

DescribeDBInstancesError

DescribeDBParameterGroupsError

DescribeDBParametersError

DescribeDBSubnetGroupsError

DescribeEngineDefaultClusterParametersError

DescribeEngineDefaultParametersError

DescribeEventCategoriesError

DescribeEventSubscriptionsError

DescribeEventsError

DescribeOrderableDBInstanceOptionsError

DescribePendingMaintenanceActionsError

DescribeValidDBInstanceModificationsError

DomainNotFoundFault

EventSubscriptionQuotaExceededFault

You have exceeded the number of events you can subscribe to.

FailoverDBClusterError

InstanceQuotaExceededFault

Request would result in user exceeding the allowed number of DB instances.

InsufficientDbClusterCapacityFault

The DB cluster does not have enough capacity for the current operation.

InsufficientDbInstanceCapacityFault

Specified DB instance class is not available in the specified Availability Zone.

InsufficientStorageClusterCapacityFault

There is insufficient storage available for the current action. You may

InvalidDbClusterEndpointStateFault

The requested operation cannot be performed on the endpoint while the endpoint is in this state.

InvalidDbClusterSnapshotStateFault

The supplied value is not a valid DB cluster snapshot state.

InvalidDbClusterStateFault

The DB cluster is not in a valid state.

InvalidDbInstanceStateFault

The specified DB instance is not in the available state.

InvalidDbParameterGroupStateFault

The DB parameter group is in use or is in an invalid state. If you are attempting to

InvalidDbSecurityGroupStateFault

The state of the DB security group does not allow deletion.

InvalidDbSnapshotStateFault

The state of the DB snapshot does not allow deletion.

InvalidDbSubnetGroupStateFault

The DB subnet group cannot be deleted because it is in use.

InvalidDbSubnetStateFault

The DB subnet is not in the available state.

InvalidEventSubscriptionStateFault

The event subscription is in an invalid state.

InvalidRestoreFault

Cannot restore from vpc backup to non-vpc DB instance.

InvalidSubnet

The requested subnet is invalid, or multiple subnets were requested that are

InvalidVpcNetworkStateFault

DB subnet group does not cover all Availability Zones after it is created

KmsKeyNotAccessibleFault

Error accessing KMS key.

ListTagsForResourceError

ModifyDBClusterEndpointError

ModifyDBClusterError

ModifyDBClusterParameterGroupError

ModifyDBClusterSnapshotAttributeError

ModifyDBInstanceError

ModifyDBParameterGroupError

ModifyDBSubnetGroupError

ModifyEventSubscriptionError

OptionGroupNotFoundFault

The designated option group could not be found.

PromoteReadReplicaDBClusterError

ProvisionedIopsNotAvailableInAzFault

Provisioned IOPS not available in the specified Availability Zone.

RebootDBInstanceError

RemoveRoleFromDBClusterError

RemoveSourceIdentifierFromSubscriptionError

RemoveTagsFromResourceError

ResetDBClusterParameterGroupError

ResetDBParameterGroupError

ResourceNotFoundFault

The specified resource ID was not found.

RestoreDBClusterFromSnapshotError

RestoreDBClusterToPointInTimeError

SharedSnapshotQuotaExceededFault

You have exceeded the maximum number of accounts that you can share a manual DB snapshot with.

SnapshotQuotaExceededFault

Request would result in user exceeding the allowed number of DB snapshots.

SnsInvalidTopicFault

The SNS topic is invalid.

SnsNoAuthorizationFault

There is no SNS authorization.

SnsTopicArnNotFoundFault

The ARN of the SNS topic could not be found.

SourceNotFoundFault

The source could not be found.

StartDBClusterError

StopDBClusterError

StorageQuotaExceededFault

Request would result in user exceeding the allowed amount of storage available across all DB instances.

StorageTypeNotSupportedFault

SubnetAlreadyInUse

The DB subnet is already in use in the Availability Zone.

SubscriptionAlreadyExistFault

This subscription already exists.

SubscriptionCategoryNotFoundFault

The designated subscription category could not be found.

SubscriptionNotFoundFault

The designated subscription could not be found.

Enums

AddRoleToDBClusterErrorKind

AddSourceIdentifierToSubscriptionErrorKind

AddTagsToResourceErrorKind

ApplyPendingMaintenanceActionErrorKind

CopyDBClusterParameterGroupErrorKind

CopyDBClusterSnapshotErrorKind

CopyDBParameterGroupErrorKind

CreateDBClusterEndpointErrorKind

CreateDBClusterErrorKind

CreateDBClusterParameterGroupErrorKind

CreateDBClusterSnapshotErrorKind

CreateDBInstanceErrorKind

CreateDBParameterGroupErrorKind

CreateDBSubnetGroupErrorKind

CreateEventSubscriptionErrorKind

DeleteDBClusterEndpointErrorKind

DeleteDBClusterErrorKind

DeleteDBClusterParameterGroupErrorKind

DeleteDBClusterSnapshotErrorKind

DeleteDBInstanceErrorKind

DeleteDBParameterGroupErrorKind

DeleteDBSubnetGroupErrorKind

DeleteEventSubscriptionErrorKind

DescribeDBClusterEndpointsErrorKind

DescribeDBClusterParameterGroupsErrorKind

DescribeDBClusterParametersErrorKind

DescribeDBClusterSnapshotAttributesErrorKind

DescribeDBClusterSnapshotsErrorKind

DescribeDBClustersErrorKind

DescribeDBEngineVersionsErrorKind

DescribeDBInstancesErrorKind

DescribeDBParameterGroupsErrorKind

DescribeDBParametersErrorKind

DescribeDBSubnetGroupsErrorKind

DescribeEngineDefaultClusterParametersErrorKind

DescribeEngineDefaultParametersErrorKind

DescribeEventCategoriesErrorKind

DescribeEventSubscriptionsErrorKind

DescribeEventsErrorKind

DescribeOrderableDBInstanceOptionsErrorKind

DescribePendingMaintenanceActionsErrorKind

DescribeValidDBInstanceModificationsErrorKind

FailoverDBClusterErrorKind

ListTagsForResourceErrorKind

ModifyDBClusterEndpointErrorKind

ModifyDBClusterErrorKind

ModifyDBClusterParameterGroupErrorKind

ModifyDBClusterSnapshotAttributeErrorKind

ModifyDBInstanceErrorKind

ModifyDBParameterGroupErrorKind

ModifyDBSubnetGroupErrorKind

ModifyEventSubscriptionErrorKind

PromoteReadReplicaDBClusterErrorKind

RebootDBInstanceErrorKind

RemoveRoleFromDBClusterErrorKind

RemoveSourceIdentifierFromSubscriptionErrorKind

RemoveTagsFromResourceErrorKind

ResetDBClusterParameterGroupErrorKind

ResetDBParameterGroupErrorKind

RestoreDBClusterFromSnapshotErrorKind

RestoreDBClusterToPointInTimeErrorKind

StartDBClusterErrorKind

StopDBClusterErrorKind