Struct aws_sdk_neptune::input::create_db_cluster_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateDbClusterInput

Implementations

impl Builder

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn character_set_name(self, input: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_character_set_name(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

If set to true, tags are copied to any snapshot of the DB cluster that is created.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon Neptune will not create a database in the DB cluster you are creating.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn db_cluster_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

pub fn set_db_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this DB cluster.

Valid Values: neptune

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to use for the new DB cluster.

Example: 1.0.2.1

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the instances in the DB cluster accept connections.

Default: 8182

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn replication_source_identifier(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

pub fn set_replication_source_identifier(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB cluster is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon Neptune will use the encryption key used to encrypt the source. Otherwise, Amazon Neptune will use your default encryption key.

If the StorageEncrypted parameter is true and ReplicationSourceIdentifier is not specified, then Amazon Neptune will use your default encryption key.

Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

If you create a Read Replica of an encrypted DB cluster in another Amazon Region, you must set KmsKeyId to a KMS key ID that is valid in the destination Amazon Region. This key is used to encrypt the Read Replica in that Amazon Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn pre_signed_url(self, input: impl Into<String>) -> Self

This parameter is not currently supported.

pub fn set_pre_signed_url(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

If set to true, enables Amazon Identity and Access Management (IAM) authentication for the entire DB cluster (this cannot be set at an instance level).

Default: false.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is enabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<CreateDbClusterInput, BuildError>

Consumes the builder and constructs a CreateDbClusterInput