Struct aws_sdk_neptune::input::CreateDbInstanceInput

#[non_exhaustive]
pub struct CreateDbInstanceInput {

    pub db_name: Option<String>,
    pub db_instance_identifier: Option<String>,
    pub allocated_storage: Option<i32>,
    pub db_instance_class: Option<String>,
    pub engine: Option<String>,
    pub master_username: Option<String>,
    pub master_user_password: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub vpc_security_group_ids: Option<Vec<String>>,
    pub availability_zone: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub backup_retention_period: Option<i32>,
    pub preferred_backup_window: Option<String>,
    pub port: Option<i32>,
    pub multi_az: Option<bool>,
    pub engine_version: Option<String>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub license_model: Option<String>,
    pub iops: Option<i32>,
    pub option_group_name: Option<String>,
    pub character_set_name: Option<String>,
    pub publicly_accessible: Option<bool>,
    pub tags: Option<Vec<Tag>>,
    pub db_cluster_identifier: Option<String>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub storage_encrypted: Option<bool>,
    pub kms_key_id: Option<String>,
    pub domain: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub monitoring_interval: Option<i32>,
    pub monitoring_role_arn: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub promotion_tier: Option<i32>,
    pub timezone: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_performance_insights: Option<bool>,
    pub performance_insights_kms_key_id: Option<String>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub deletion_protection: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

db_name: Option<String>

Not supported.

db_instance_identifier: Option<String>

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

allocated_storage: Option<i32>

Not supported by Neptune.

db_instance_class: Option<String>

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Regions.

engine: Option<String>

The name of the database engine to be used for this instance.

Valid Values: neptune

master_username: Option<String>

Not supported by Neptune.

master_user_password: Option<String>

Not supported by Neptune.

db_security_groups: Option<Vec<String>>

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB instance.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

availability_zone: Option<String>

The EC2 Availability Zone that the DB instance is created in

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same Amazon Region as the current endpoint.

db_subnet_group_name: Option<String>

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

preferred_maintenance_window: Option<String>

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

backup_retention_period: Option<i32>

The number of days for which automated backups are retained.

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

preferred_backup_window: Option<String>

The daily time range during which automated backups are created.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

port: Option<i32>

The port number on which the database accepts connections.

Not applicable. The port is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 8182

Type: Integer

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

engine_version: Option<String>

The version number of the database engine to use. Currently, setting this parameter has no effect.

auto_minor_version_upgrade: Option<bool>

Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

license_model: Option<String>

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

iops: Option<i32>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

option_group_name: Option<String>

(Not supported by Neptune)

character_set_name: Option<String>

(Not supported by Neptune)

publicly_accessible: Option<bool>

This flag should no longer be used.

tags: Option<Vec<Tag>>

The tags to assign to the new instance.

db_cluster_identifier: Option<String>

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Not applicable. Storage is managed by the DB Cluster.

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

storage_encrypted: Option<bool>

Specifies whether the DB instance is encrypted.

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

kms_key_id: Option<String>

The Amazon KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same Amazon account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon Neptune will use your default encryption key. Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

domain: Option<String>

Specify the Active Directory Domain to create the instance in.

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

monitoring_interval: Option<i32>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

promotion_tier: Option<i32>

A value that specifies the order in which an Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

timezone: Option<String>

The time zone of the DB instance.

enable_iam_database_authentication: Option<bool>

Not supported by Neptune (ignored).

enable_performance_insights: Option<bool>

(Not supported by Neptune)

performance_insights_kms_key_id: Option<String>

(Not supported by Neptune)

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of log types that need to be enabled for exporting to CloudWatch Logs.

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

DB instances in a DB cluster can be deleted even when deletion protection is enabled in their parent DB cluster.

Implementations

impl CreateDbInstanceInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateDBInstance, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateDBInstance>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateDbInstanceInput