Struct aws_sdk_neptune::input::ModifyDbInstanceInput

#[non_exhaustive]
pub struct ModifyDbInstanceInput {

    pub db_instance_identifier: Option<String>,
    pub allocated_storage: Option<i32>,
    pub db_instance_class: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub db_security_groups: Option<Vec<String>>,
    pub vpc_security_group_ids: Option<Vec<String>>,
    pub apply_immediately: bool,
    pub master_user_password: Option<String>,
    pub db_parameter_group_name: Option<String>,
    pub backup_retention_period: Option<i32>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub multi_az: Option<bool>,
    pub engine_version: Option<String>,
    pub allow_major_version_upgrade: bool,
    pub auto_minor_version_upgrade: Option<bool>,
    pub license_model: Option<String>,
    pub iops: Option<i32>,
    pub option_group_name: Option<String>,
    pub new_db_instance_identifier: Option<String>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub ca_certificate_identifier: Option<String>,
    pub domain: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub monitoring_interval: Option<i32>,
    pub db_port_number: Option<i32>,
    pub publicly_accessible: Option<bool>,
    pub monitoring_role_arn: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub promotion_tier: Option<i32>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_performance_insights: Option<bool>,
    pub performance_insights_kms_key_id: Option<String>,
    pub cloudwatch_logs_export_configuration: Option<CloudwatchLogsExportConfiguration>,
    pub deletion_protection: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

db_instance_identifier: Option<String>

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

allocated_storage: Option<i32>

Not supported by Neptune.

db_instance_class: Option<String>

The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting

db_subnet_group_name: Option<String>

The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you specify true for the ApplyImmediately parameter.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

db_security_groups: Option<Vec<String>>

A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

apply_immediately: bool

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance.

If this parameter is set to false, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot.

Default: false

master_user_password: Option<String>

Not supported by Neptune.

db_parameter_group_name: Option<String>

The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. The db instance will NOT be rebooted automatically and the parameter changes will NOT be applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

backup_retention_period: Option<i32>

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

preferred_backup_window: Option<String>

The daily time range during which automated backups are created if automated backups are enabled.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

preferred_maintenance_window: Option<String>

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

multi_az: Option<bool>

Specifies if the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

engine_version: Option<String>

The version number of the database engine to upgrade to. Currently, setting this parameter has no effect. To upgrade your database engine to the most recent release, use the ApplyPendingMaintenanceAction API.

allow_major_version_upgrade: bool

Indicates that major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

auto_minor_version_upgrade: Option<bool>

Indicates that minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and Neptune has enabled auto patching for that engine version.

license_model: Option<String>

Not supported by Neptune.

iops: Option<i32>

The new Provisioned IOPS (I/O operations per second) value for the instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

Default: Uses existing setting

option_group_name: Option<String>

(Not supported by Neptune)

new_db_instance_identifier: Option<String>

The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot will occur immediately if you set Apply Immediately to true, or will occur during the next maintenance window if Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

storage_type: Option<String>

Not supported.

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

ca_certificate_identifier: Option<String>

Indicates the certificate that needs to be associated with the instance.

domain: Option<String>

Not supported.

copy_tags_to_snapshot: Option<bool>

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

monitoring_interval: Option<i32>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

db_port_number: Option<i32>

The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

Default: 8182

publicly_accessible: Option<bool>

This flag should no longer be used.

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

domain_iam_role_name: Option<String>

Not supported

promotion_tier: Option<i32>

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

enable_iam_database_authentication: Option<bool>

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines

Not applicable. Mapping Amazon IAM accounts to database accounts is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: false

enable_performance_insights: Option<bool>

(Not supported by Neptune)

performance_insights_kms_key_id: Option<String>

(Not supported by Neptune)

cloudwatch_logs_export_configuration: Option<CloudwatchLogsExportConfiguration>

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance or DB cluster.

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

Implementations

impl ModifyDbInstanceInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<ModifyDBInstance, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<ModifyDBInstance>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ModifyDbInstanceInput