Struct aws_sdk_neptune::model::db_cluster_snapshot::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbClusterSnapshot

Implementations

impl Builder

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn db_cluster_snapshot_identifier(self, input: impl Into<String>) -> Self

Specifies the identifier for a DB cluster snapshot. Must match the identifier of an existing snapshot.

After you restore a DB cluster using a DBClusterSnapshotIdentifier, you must specify the same DBClusterSnapshotIdentifier for any future updates to the DB cluster. When you specify this property for an update, the DB cluster is not restored from the snapshot again, and the data in the database is not changed.

However, if you don't specify the DBClusterSnapshotIdentifier, an empty DB cluster is created, and the original DB cluster is deleted. If you specify a property that is different from the previous snapshot restore property, the DB cluster is restored from the snapshot specified by the DBClusterSnapshotIdentifier, and the original DB cluster is deleted.

pub fn set_db_cluster_snapshot_identifier(self, input: Option<String>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

Specifies the DB cluster identifier of the DB cluster that this DB cluster snapshot was created from.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn snapshot_create_time(self, input: Instant) -> Self

Provides the time when the snapshot was taken, in Universal Coordinated Time (UTC).

pub fn set_snapshot_create_time(self, input: Option<Instant>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

Specifies the name of the database engine.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

Specifies the allocated storage size in gibibytes (GiB).

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

Specifies the status of this DB cluster snapshot.

pub fn set_status(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

Specifies the port that the DB cluster was listening on at the time of the snapshot.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

Provides the VPC ID associated with the DB cluster snapshot.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn cluster_create_time(self, input: Instant) -> Self

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

pub fn set_cluster_create_time(self, input: Option<Instant>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

Provides the version of the database engine for this DB cluster snapshot.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

Provides the license model information for this DB cluster snapshot.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn snapshot_type(self, input: impl Into<String>) -> Self

Provides the type of the DB cluster snapshot.

pub fn set_snapshot_type(self, input: Option<String>) -> Self

pub fn percent_progress(self, input: i32) -> Self

Specifies the percentage of the estimated data that has been transferred.

pub fn set_percent_progress(self, input: Option<i32>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB cluster snapshot is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster snapshot.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn db_cluster_snapshot_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB cluster snapshot.

pub fn set_db_cluster_snapshot_arn(self, input: Option<String>) -> Self

pub fn source_db_cluster_snapshot_arn(self, input: impl Into<String>) -> Self

If the DB cluster snapshot was copied from a source DB cluster snapshot, the Amazon Resource Name (ARN) for the source DB cluster snapshot, otherwise, a null value.

pub fn set_source_db_cluster_snapshot_arn(self, input: Option<String>) -> Self

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

pub fn build(self) -> DbClusterSnapshot

Consumes the builder and constructs a DbClusterSnapshot