Struct aws_sdk_neptune::model::DbCluster

#[non_exhaustive]
pub struct DbCluster {

    pub allocated_storage: Option<i32>,
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i32>,
    pub character_set_name: Option<String>,
    pub database_name: Option<String>,
    pub db_cluster_identifier: Option<String>,
    pub db_cluster_parameter_group: Option<String>,
    pub db_subnet_group: Option<String>,
    pub status: Option<String>,
    pub percent_progress: Option<String>,
    pub earliest_restorable_time: Option<Instant>,
    pub endpoint: Option<String>,
    pub reader_endpoint: Option<String>,
    pub multi_az: bool,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub latest_restorable_time: Option<Instant>,
    pub port: Option<i32>,
    pub master_username: Option<String>,
    pub db_cluster_option_group_memberships: Option<Vec<DbClusterOptionGroupStatus>>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_source_identifier: Option<String>,
    pub read_replica_identifiers: Option<Vec<String>>,
    pub db_cluster_members: Option<Vec<DbClusterMember>>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
    pub hosted_zone_id: Option<String>,
    pub storage_encrypted: bool,
    pub kms_key_id: Option<String>,
    pub db_cluster_resource_id: Option<String>,
    pub db_cluster_arn: Option<String>,
    pub associated_roles: Option<Vec<DbClusterRole>>,
    pub iam_database_authentication_enabled: bool,
    pub clone_group_id: Option<String>,
    pub cluster_create_time: Option<Instant>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub enabled_cloudwatch_logs_exports: Option<Vec<String>>,
    pub deletion_protection: Option<bool>,
    pub cross_account_clone: Option<bool>,
    pub automatic_restart_time: Option<Instant>,

}

Contains the details of an Amazon Neptune DB cluster.

This data type is used as a response element in the DescribeDBClusters action.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

allocated_storage: Option<i32>

AllocatedStorage always returns 1, because Neptune DB cluster storage size is not fixed, but instead automatically adjusts as needed.

availability_zones: Option<Vec<String>>

Provides the list of EC2 Availability Zones that instances in the DB cluster can be created in.

backup_retention_period: Option<i32>

Specifies the number of days for which automatic DB snapshots are retained.

character_set_name: Option<String>

Not supported by Neptune.

database_name: Option<String>

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

db_cluster_identifier: Option<String>

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

db_cluster_parameter_group: Option<String>

Specifies the name of the DB cluster parameter group for the DB cluster.

db_subnet_group: Option<String>

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

status: Option<String>

Specifies the current state of this DB cluster.

percent_progress: Option<String>

Specifies the progress of the operation as a percentage.

earliest_restorable_time: Option<Instant>

Specifies the earliest time to which a database can be restored with point-in-time restore.

endpoint: Option<String>

Specifies the connection endpoint for the primary instance of the DB cluster.

reader_endpoint: Option<String>

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Read Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Neptune distributes the connection requests among the Read Replicas in the DB cluster. This functionality can help balance your read workload across multiple Read Replicas in your DB cluster.

If a failover occurs, and the Read Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Read Replicas in the cluster, you can then reconnect to the reader endpoint.

multi_az: bool

Specifies whether the DB cluster has instances in multiple Availability Zones.

engine: Option<String>

Provides the name of the database engine to be used for this DB cluster.

engine_version: Option<String>

Indicates the database engine version.

latest_restorable_time: Option<Instant>

Specifies the latest time to which a database can be restored with point-in-time restore.

port: Option<i32>

Specifies the port that the database engine is listening on.

master_username: Option<String>

Not supported by Neptune.

db_cluster_option_group_memberships: Option<Vec<DbClusterOptionGroupStatus>>

Not supported by Neptune.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

replication_source_identifier: Option<String>

Not supported by Neptune.

read_replica_identifiers: Option<Vec<String>>

Contains one or more identifiers of the Read Replicas associated with this DB cluster.

db_cluster_members: Option<Vec<DbClusterMember>>

Provides the list of instances that make up the DB cluster.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of VPC security groups that the DB cluster belongs to.

hosted_zone_id: Option<String>

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

storage_encrypted: bool

Specifies whether the DB cluster is encrypted.

kms_key_id: Option<String>

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster.

db_cluster_resource_id: Option<String>

The Amazon Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB cluster is accessed.

db_cluster_arn: Option<String>

The Amazon Resource Name (ARN) for the DB cluster.

associated_roles: Option<Vec<DbClusterRole>>

Provides a list of the Amazon Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon services on your behalf.

iam_database_authentication_enabled: bool

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

clone_group_id: Option<String>

Identifies the clone group to which the DB cluster is associated.

cluster_create_time: Option<Instant>

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

copy_tags_to_snapshot: Option<bool>

If set to true, tags are copied to any snapshot of the DB cluster that is created.

enabled_cloudwatch_logs_exports: Option<Vec<String>>

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

deletion_protection: Option<bool>

Indicates whether or not the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

cross_account_clone: Option<bool>

If set to true, the DB cluster can be cloned across accounts.

automatic_restart_time: Option<Instant>

Time at which the DB cluster will be automatically restarted.

Implementations

impl DbCluster

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DbCluster