Module aws_sdk_neptune::operation

Structs

AddRoleToDBCluster

Associates an Identity and Access Management (IAM) role from an

AddSourceIdentifierToSubscription

Adds a source identifier to an existing event notification subscription.

AddTagsToResource

Adds metadata tags to an Amazon Neptune resource. These tags can also be used with cost

ApplyPendingMaintenanceAction

Applies a pending maintenance action to a resource (for example, to a DB instance).

CopyDBClusterParameterGroup

Copies the specified DB cluster parameter group.

CopyDBClusterSnapshot

Copies a snapshot of a DB cluster.

CopyDBParameterGroup

Copies the specified DB parameter group.

CreateDBCluster

Creates a new Amazon Neptune DB cluster.

CreateDBClusterEndpoint

Creates a new custom endpoint and associates it with an Amazon Neptune DB cluster.

CreateDBClusterParameterGroup

Creates a new DB cluster parameter group.

CreateDBClusterSnapshot

Creates a snapshot of a DB cluster.

CreateDBInstance

Creates a new DB instance.

CreateDBParameterGroup

Creates a new DB parameter group.

CreateDBSubnetGroup

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at

CreateEventSubscription

Creates an event notification subscription. This action requires a topic ARN (Amazon

DeleteDBCluster

The DeleteDBCluster action deletes a previously provisioned DB cluster. When you delete a

DeleteDBClusterEndpoint

Deletes a custom endpoint and removes it from an Amazon Neptune DB cluster.

DeleteDBClusterParameterGroup

Deletes a specified DB cluster parameter group. The DB cluster parameter group to be

DeleteDBClusterSnapshot

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is

DeleteDBInstance

The DeleteDBInstance action deletes a previously provisioned DB instance. When you delete

DeleteDBParameterGroup

Deletes a specified DBParameterGroup. The DBParameterGroup to be deleted can't be

DeleteDBSubnetGroup

Deletes a DB subnet group.

DeleteEventSubscription

Deletes an event notification subscription.

DescribeDBClusterEndpoints

Returns information about endpoints for an Amazon Neptune DB cluster.

DescribeDBClusterParameterGroups

Returns a list of DBClusterParameterGroup descriptions. If a

DescribeDBClusterParameters

Returns the detailed parameter list for a particular DB cluster parameter group.

DescribeDBClusterSnapshotAttributes

Returns a list of DB cluster snapshot attribute names and values for a manual DB cluster

DescribeDBClusterSnapshots

Returns information about DB cluster snapshots. This API action supports

DescribeDBClusters

Returns information about provisioned DB clusters, and supports

DescribeDBEngineVersions

Returns a list of the available DB engines.

DescribeDBInstances

Returns information about provisioned instances, and supports pagination.

DescribeDBParameterGroups

Returns a list of DBParameterGroup descriptions. If a

DescribeDBParameters

Returns the detailed parameter list for a particular DB parameter group.

DescribeDBSubnetGroups

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the

DescribeEngineDefaultClusterParameters

Returns the default engine and system parameter information for the cluster database

DescribeEngineDefaultParameters

Returns the default engine and system parameter information for the specified database

DescribeEventCategories

Displays a list of categories for all event source types, or, if specified, for a

DescribeEventSubscriptions

Lists all the subscription descriptions for a customer account. The description for a

DescribeEvents

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter

DescribeOrderableDBInstanceOptions

Returns a list of orderable DB instance options for the specified engine.

DescribePendingMaintenanceActions

Returns a list of resources (for example, DB instances) that have at least one pending

DescribeValidDBInstanceModifications

You can call DescribeValidDBInstanceModifications

FailoverDBCluster

Forces a failover for a DB cluster.

ListTagsForResource

Lists all tags on an Amazon Neptune resource.

ModifyDBCluster

Modify a setting for a DB cluster. You can change one or more database configuration

ModifyDBClusterEndpoint

Modifies the properties of an endpoint in an Amazon Neptune DB cluster.

ModifyDBClusterParameterGroup

Modifies the parameters of a DB cluster parameter group. To modify more than one

ModifyDBClusterSnapshotAttribute

Adds an attribute and values to, or removes an attribute and values from, a manual DB

ModifyDBInstance

Modifies settings for a DB instance. You can change one or more database configuration

ModifyDBParameterGroup

Modifies the parameters of a DB parameter group. To modify more than one parameter,

ModifyDBSubnetGroup

Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in

ModifyEventSubscription

Modifies an existing event notification subscription. Note that you can't modify the

PromoteReadReplicaDBCluster

Not supported.

RebootDBInstance

You might need to reboot your DB instance, usually for maintenance reasons. For example,

RemoveRoleFromDBCluster

Disassociates an Identity and Access Management (IAM) role from a DB cluster.

RemoveSourceIdentifierFromSubscription

Removes a source identifier from an existing event notification subscription.

RemoveTagsFromResource

Removes metadata tags from an Amazon Neptune resource.

ResetDBClusterParameterGroup

Modifies the parameters of a DB cluster parameter group to the default value. To reset

ResetDBParameterGroup

Modifies the parameters of a DB parameter group to the engine/system default value. To

RestoreDBClusterFromSnapshot

Creates a new DB cluster from a DB snapshot or DB cluster snapshot.

RestoreDBClusterToPointInTime

Restores a DB cluster to an arbitrary point in time. Users can restore to any point in

StartDBCluster

Starts an Amazon Neptune DB cluster that was stopped using the AWS

StopDBCluster

Stops an Amazon Neptune DB cluster. When you stop a DB cluster, Neptune