Struct aws_sdk_neptune::model::db_instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbInstance

Implementations

impl Builder

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn db_instance_class(self, input: impl Into<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn engine(self, input: impl Into<String>) -> Self

Provides the name of the database engine to be used for this DB instance.

pub fn set_engine(self, input: Option<String>) -> Self

Provides the name of the database engine to be used for this DB instance.

pub fn db_instance_status(self, input: impl Into<String>) -> Self

Specifies the current state of this database.

pub fn set_db_instance_status(self, input: Option<String>) -> Self

Specifies the current state of this database.

pub fn master_username(self, input: impl Into<String>) -> Self

Not supported by Neptune.

pub fn set_master_username(self, input: Option<String>) -> Self

Not supported by Neptune.

pub fn db_name(self, input: impl Into<String>) -> Self

The database name.

pub fn set_db_name(self, input: Option<String>) -> Self

The database name.

pub fn endpoint(self, input: Endpoint) -> Self

Specifies the connection endpoint.

pub fn set_endpoint(self, input: Option<Endpoint>) -> Self

Specifies the connection endpoint.

pub fn allocated_storage(self, input: i32) -> Self

Not supported by Neptune.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

Not supported by Neptune.

pub fn instance_create_time(self, input: Instant) -> Self

Provides the date and time the DB instance was created.

pub fn set_instance_create_time(self, input: Option<Instant>) -> Self

Provides the date and time the DB instance was created.

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn backup_retention_period(self, input: i32) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn db_security_groups(
self,
input: impl Into<DbSecurityGroupMembership>
) -> Self

Appends an item to db_security_groups.

To override the contents of this collection use set_db_security_groups.

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

pub fn set_db_security_groups(
self,
input: Option<Vec<DbSecurityGroupMembership>>
) -> Self

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

Appends an item to vpc_security_groups.

To override the contents of this collection use set_vpc_security_groups.

Provides a list of VPC security group elements that the DB instance belongs to.

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

Provides a list of VPC security group elements that the DB instance belongs to.

pub fn db_parameter_groups(
self,
input: impl Into<DbParameterGroupStatus>
) -> Self

Appends an item to db_parameter_groups.

To override the contents of this collection use set_db_parameter_groups.

Provides the list of DB parameter groups applied to this DB instance.

pub fn set_db_parameter_groups(
self,
input: Option<Vec<DbParameterGroupStatus>>
) -> Self

Provides the list of DB parameter groups applied to this DB instance.

pub fn availability_zone(self, input: impl Into<String>) -> Self

Specifies the name of the Availability Zone the DB instance is located in.

pub fn set_availability_zone(self, input: Option<String>) -> Self

Specifies the name of the Availability Zone the DB instance is located in.

pub fn db_subnet_group(self, input: DbSubnetGroup) -> Self

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<DbSubnetGroup>) -> Self

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn pending_modified_values(self, input: PendingModifiedValues) -> Self

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn set_pending_modified_values(
self,
input: Option<PendingModifiedValues>
) -> Self

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn latest_restorable_time(self, input: Instant) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn set_latest_restorable_time(self, input: Option<Instant>) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn multi_az(self, input: bool) -> Self

Specifies if the DB instance is a Multi-AZ deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

Specifies if the DB instance is a Multi-AZ deployment.

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

Indicates the database engine version.

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

Indicates that minor version patches are applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

Indicates that minor version patches are applied automatically.

pub fn read_replica_source_db_instance_identifier(
self,
input: impl Into<String>
) -> Self

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

pub fn set_read_replica_source_db_instance_identifier(
self,
input: Option<String>
) -> Self

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

pub fn read_replica_db_instance_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_instance_identifiers.

To override the contents of this collection use set_read_replica_db_instance_identifiers.

Contains one or more identifiers of the Read Replicas associated with this DB instance.

pub fn set_read_replica_db_instance_identifiers(
self,
input: Option<Vec<String>>
) -> Self

Contains one or more identifiers of the Read Replicas associated with this DB instance.

pub fn read_replica_db_cluster_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_cluster_identifiers.

To override the contents of this collection use set_read_replica_db_cluster_identifiers.

Contains one or more identifiers of DB clusters that are Read Replicas of this DB instance.

pub fn set_read_replica_db_cluster_identifiers(
self,
input: Option<Vec<String>>
) -> Self

Contains one or more identifiers of DB clusters that are Read Replicas of this DB instance.

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for this DB instance.

pub fn set_license_model(self, input: Option<String>) -> Self

License model information for this DB instance.

pub fn iops(self, input: i32) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value.

pub fn set_iops(self, input: Option<i32>) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value.

pub fn option_group_memberships(
self,
input: impl Into<OptionGroupMembership>
) -> Self

Appends an item to option_group_memberships.

To override the contents of this collection use set_option_group_memberships.

(Not supported by Neptune)

pub fn set_option_group_memberships(
self,
input: Option<Vec<OptionGroupMembership>>
) -> Self

(Not supported by Neptune)

pub fn character_set_name(self, input: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_character_set_name(self, input: Option<String>) -> Self

(Not supported by Neptune)

pub fn secondary_availability_zone(self, input: impl Into<String>) -> Self

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

pub fn set_secondary_availability_zone(self, input: Option<String>) -> Self

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

pub fn publicly_accessible(self, input: bool) -> Self

This flag should no longer be used.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

This flag should no longer be used.

pub fn status_infos(self, input: impl Into<DbInstanceStatusInfo>) -> Self

Appends an item to status_infos.

To override the contents of this collection use set_status_infos.

The status of a Read Replica. If the instance is not a Read Replica, this is blank.

pub fn set_status_infos(self, input: Option<Vec<DbInstanceStatusInfo>>) -> Self

The status of a Read Replica. If the instance is not a Read Replica, this is blank.

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type associated with DB instance.

pub fn set_storage_type(self, input: Option<String>) -> Self

Specifies the storage type associated with DB instance.

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn db_instance_port(self, input: i32) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn set_db_instance_port(self, input: Option<i32>) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn storage_encrypted(self, input: bool) -> Self

Not supported: The encryption for DB instances is managed by the DB cluster.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

Not supported: The encryption for DB instances is managed by the DB cluster.

pub fn kms_key_id(self, input: impl Into<String>) -> Self

Not supported: The encryption for DB instances is managed by the DB cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

Not supported: The encryption for DB instances is managed by the DB cluster.

pub fn dbi_resource_id(self, input: impl Into<String>) -> Self

The Amazon Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB instance is accessed.

pub fn set_dbi_resource_id(self, input: Option<String>) -> Self

The Amazon Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB instance is accessed.

pub fn ca_certificate_identifier(self, input: impl Into<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn domain_memberships(self, input: impl Into<DomainMembership>) -> Self

Appends an item to domain_memberships.

To override the contents of this collection use set_domain_memberships.

Not supported

pub fn set_domain_memberships(
self,
input: Option<Vec<DomainMembership>>
) -> Self

Not supported

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

pub fn enhanced_monitoring_resource_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

pub fn set_enhanced_monitoring_resource_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits Neptune to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

The ARN for the IAM role that permits Neptune to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

pub fn promotion_tier(self, input: i32) -> Self

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

pub fn db_instance_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB instance.

pub fn set_db_instance_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) for the DB instance.

pub fn timezone(self, input: impl Into<String>) -> Self

Not supported.

pub fn set_timezone(self, input: Option<String>) -> Self

Not supported.

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if Amazon Identity and Access Management (IAM) authentication is enabled, and otherwise false.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

True if Amazon Identity and Access Management (IAM) authentication is enabled, and otherwise false.

pub fn performance_insights_enabled(self, input: bool) -> Self

(Not supported by Neptune)

pub fn set_performance_insights_enabled(self, input: Option<bool>) -> Self

(Not supported by Neptune)

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

(Not supported by Neptune)

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

(Not supported by Neptune)

pub fn enabled_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

Appends an item to enabled_cloudwatch_logs_exports.

To override the contents of this collection use set_enabled_cloudwatch_logs_exports.

A list of log types that this DB instance is configured to export to CloudWatch Logs.

pub fn set_enabled_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

A list of log types that this DB instance is configured to export to CloudWatch Logs.

pub fn deletion_protection(self, input: bool) -> Self

Indicates whether or not the DB instance has deletion protection enabled. The instance can't be deleted when deletion protection is enabled. See Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

Indicates whether or not the DB instance has deletion protection enabled. The instance can't be deleted when deletion protection is enabled. See Deleting a DB Instance.

pub fn build(self) -> DbInstance

Consumes the builder and constructs a DbInstance